Let LD comunity flag posts as spam now.
This code looks for the last 100 comments in a post. If one comment contains "SPAMREPORT" or "SPAM REPORT", then the post is hidden. This check is only performed if the post is not already flagged by the previously coded rules. The 'gamecount' rule has been removed by me, as it is not really useful anymore seeing the bots acctually posting bogus-games now. (e.g. the cocainebear one).